estate_planning_backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
django-cors-headers>=4.0.0
djangorestframework-simplejwt>=5.2.2
django-filter>=23.2
Pillow>=9.5.0
openai>=0.27.0
gunicorn>=20.1.0
whitenoise>=6.4.0
dj-database-url>=1.0.0
pytest>=7.3.1
pytest-django>=4.5.2
pytest-cov>=4.0.0 